+++ calculator/src/main/java/ru/chainichek/neostudy/calculator/model/EmploymentPosition.java
@@ -1,7 +1,8 @@
 package ru.chainichek.neostudy.calculator.model;
 
-public enum Position {
+public enum EmploymentPosition {
     MIDDLE_MANAGER,  // Менеджер среднего звена
     TOP_MANAGER,  // Топ-менеджер
-    OTHER  // Всё остальное
+    WORKER,
+    OWNER   // Всё остальное
 }