+++ deal/src/main/java/ru/chainichek/neostudy/deal/dto/statement/EmploymentDto.java
@@ -1,5 +1,6 @@
 package ru.chainichek.neostudy.deal.dto.statement;
 
+import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.NotNull;
 import ru.chainichek.neostudy.deal.model.client.EmploymentStatus;
 import ru.chainichek.neostudy.deal.model.client.EmploymentPosition;
@@ -9,13 +10,17 @@ import java.math.BigDecimal;
 public record EmploymentDto(@NotNull
                             EmploymentStatus employmentStatus,
                             @NotNull
+                            @Schema(example = "251112724508")
                             String employerINN,
                             @NotNull
+                            @Schema(example = "300000")
                             BigDecimal salary,
                             @NotNull
                             EmploymentPosition position,
                             @NotNull
+                            @Schema(example = "18")
                             Integer workExperienceTotal,
                             @NotNull
+                            @Schema(example = "3")
                             Integer workExperienceCurrent) {
 }