+++ deal/src/main/java/ru/chainichek/neostudy/deal/model/dossier/EmailTheme.java
@@ -1,0 +1,10 @@
+package ru.chainichek.neostudy.deal.model.dossier;
+
+public enum EmailTheme {
+    FINISH_REGISTRATION,
+    CREATE_DOCUMENTS,
+    SEND_DOCUMENTS,
+    SEND_SES,
+    CREDIT_ISSUED,
+    STATEMENT_DENIED
+}