+++ calculator/CHANGELOG.md
@@ -1,3 +1,8 @@
 # Change log
 ## 1.0.0
 * MVP-1 microservice 'Calculator'
+## 1.0.1
+* Added k8s support
+* Fixed missing enum values from MS 'Deal'
+* Removed mvnw files
+* Added base application.properties file